Implement the delete operation and add a MutableInteger implementation
The delete operation supports aborting and tracks the deletion progress. The Analyzer class was extended to support
individual aborting of analysis or deletion.

The MutableInteger class is useful when the state need to change, but the integers are not suitable as class fields as
it was the case for the Analyzer class and the integer field "totalTargets".